# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Watchlist API",
    "version": "1.0.0",
    "description": "API documentation for Movie Watchlist project"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/movies": {
      "get": {
        "summary": "Get all movies",
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "Movies fetched successfully"
          }
        }
      },
      "post": {
        "summary": "Create a new movie",
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "genre",
                  "releaseYear",
                  "description"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Inception"
                  },
                  "genre": {
                    "type": "string",
                    "example": "Sci-Fi"
                  },
                  "releaseYear": {
                    "type": "integer",
                    "example": 2010
                  },
                  "description": {
                    "type": "string",
                    "example": "A mind-bending thriller"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movie created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/movies/search-tmdb": {
      "get": {
        "summary": "Search movies from TMDB by title",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Inception"
          }
        ],
        "responses": {
          "200": {
            "description": "TMDB movies fetched successfully"
          },
          "400": {
            "description": "Missing query parameter"
          },
          "500": {
            "description": "Failed to search TMDB"
          }
        }
      }
    },
    "/movies/{id}": {
      "get": {
        "summary": "Get movie by ID",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie fetched successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Movie not found"
          }
        }
      },
      "put": {
        "summary": "Update a movie",
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Interstellar"
                  },
                  "genre": {
                    "type": "string",
                    "example": "Sci-Fi"
                  },
                  "releaseYear": {
                    "type": "integer",
                    "example": 2014
                  },
                  "description": {
                    "type": "string",
                    "example": "A science fiction film about space and time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Movie updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Movie not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a movie",
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie deleted successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Movie not found"
          }
        }
      }
    },
    "/reviews": {
      "get": {
        "summary": "Get all reviews",
        "tags": [
          "Reviews"
        ],
        "responses": {
          "200": {
            "description": "Reviews fetched successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "movieId",
                  "rating"
                ],
                "properties": {
                  "movieId": {
                    "type": "string",
                    "example": "movie001"
                  },
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "comment": {
                    "type": "string",
                    "example": "Amazing movie"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/reviews/{id}": {
      "get": {
        "summary": "Get review by ID",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review fetched successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "put": {
        "summary": "Update a review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 4
                  },
                  "comment": {
                    "type": "string",
                    "example": "Updated review"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Review not found"
          }
        }
      }
    },
    "/watchlists": {
      "get": {
        "summary": "Get all watchlists",
        "tags": [
          "Watchlists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Watchlists fetched successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a watchlist entry",
        "tags": [
          "Watchlists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "user001"
                  },
                  "movieId": {
                    "type": "string",
                    "example": "movie001"
                  },
                  "status": {
                    "type": "string",
                    "example": "watching"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Watchlist created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/watchlists/{id}": {
      "put": {
        "summary": "Update a watchlist entry",
        "tags": [
          "Watchlists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watchlist updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Watchlist not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a watchlist entry",
        "tags": [
          "Watchlists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watchlist deleted successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Watchlist not found"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Movies",
      "description": "Movie management endpoints"
    },
    {
      "name": "Reviews",
      "description": "Review management endpoints"
    },
    {
      "name": "Watchlists",
      "description": "Watchlist management endpoints"
    }
  ]
}